Add Channel Preview golden test
