+++ .gitignore
@@ -62,3 +62,6 @@ doc/api/
 fvm
 google-services.json
 example/ios/dist
+
+# don't check in golden failure output
+**/failures/*.png
