+++ pubspec.yaml
@@ -28,8 +28,16 @@ dependencies:
   visibility_detector: ^0.1.5
   http_parser: ^3.1.4
 
+flutter:
+  # Needed for Material icons like Icons.battery to be rendered in goldens
+  uses-material-design: true
+  # https://github.com/eBay/flutter_glove_box/issues/31
+  assets:
+    - images/
+
 dev_dependencies:
   pedantic: ^1.9.0
   flutter_test:
     sdk: flutter
   mockito: ^4.1.1
+  golden_toolkit: ^0.6.0
