+++ test/flutter_test_config.dart
@@ -1,0 +1,8 @@
+import 'dart:async';
+
+import 'package:golden_toolkit/golden_toolkit.dart';
+
+Future<void> main(FutureOr<void> Function() testMain) async {
+  await loadAppFonts();
+  return testMain();
+}
